Added more properties for zipped packages and enhanced script sequence and version to support Long type for longer run.

+++ src/main/java/com/techyplanet/scriptdeployer/repository/ScriptHistoryRepository.java
@@ -6,7 +6,7 @@ import com.techyplanet.scriptdeployer.entity.ScriptHistory;
 
 public interface ScriptHistoryRepository extends CrudRepository<ScriptHistory, Long> {
 
-	ScriptHistory findBySequenceAndType(Integer sequence, String type);
+	ScriptHistory findBySequenceAndType(Long sequence, String type);
 
 	ScriptHistory findByPath(String path);
 }
